Alo_LeafDisease/requirements.txt
streamlit>=1.29.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-google-genai>=0.0.6
langchain-core>=0.1.23
python-dotenv>=1.0.0
pandas>=2.0.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

